week 4/marketing-insights-analyzer/requirements.txt
openai>=1.3.0
python-dotenv
pandas
